Move view into components. #11
