+++ config/application.rb
@@ -21,5 +21,8 @@ module Minesweeper
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
+    #
+    config.browserify_rails.source_map_environments << 'development'
+    config.browserify_rails.commandline_options = "-t vueify --extension=\".js.vue\""
   end
 end
