+++ package.json
@@ -7,7 +7,9 @@
     "axios": "^0.15.3",
     "browserify": "~10.2.4",
     "browserify-incremental": "^3.0.1",
-    "vue": "^2.1.10"
+    "vue": "^2.1.10",
+    "vue-router": "^2.2.0",
+    "vueify": "^9.4.0"
   },
   "license": "MIT",
   "engines": {
